Automate setup with make
This commit moves all the documented setup code into a make target so
that it can be executed by a computer instead of a human.

This means that setting up the project to work on now looks like the
following:

```sh
git clone --recursive git@github.com:cds-snc/tracker.git
cd tracker
make setup
. .env/bin/activate
tracker run
```

It's assumed that your computer has pyenv installed so that the python
version specified in .python-version is respected.