Move to monorepo layout
This commit combines the projects formerly known as tracker and
track-web into a single monorepo.
As part of this change, and in light of the fact that the project in
it's entirety is referred to as tracker, the contents of the tracker
repo is now kept in the scanner subdirectory, and track-web is now found
in the web directory.

The resulting structure is:

```sh
.
├── platform
├── README.md
├── scanner
└── web
```