Add docker-compose based frontend dev workflow
This commit add docker-compose to create a nice frontend dev workflow.
Docker compose starts the frontend service with a bind-mount for the current
directory along with a mocked graphql api.
To knit those together Envoy is used as a lightweight proxy allowing the two
services to appear as the same domain.

Envoy was chosen here because it's lightweight and fast and also because it's
already included in the project by virtue of being used by Istio.