Add docker-compose for api development
This commit adds a docker-compose file to help make a nice dev workflow.

In the api folder, you can now run the command `docker-compose up` and have the
api and database started and listening on port 5000.

The file is also set up so that it runs the migrations before starting the api.

If you want to run everything in the background, you can "detach" like this:

```sh
docker-compose up -d
```

Stopping everything is basically what you might expect:

```sh
docker-compose down
```