Update to @testing-library/react 10
This caused lots of warnings in our tests so I went through all of them
updating to the use the new functions and did some other refactoring as I went.

Notable side effects of this refactoring:

I was able to drop the user-event
library entirely in favor of fireEvent. All of the tests that used used
user-event would run and never complete.

By default the QR code library renders a canvas element. Supporting these in
JSDOM require another library which compiles things and so I switched the QR
Code to use SVG instead. Better accessibility and easier to test.

I realised that the QRCodePage wasn't really testing anything, so I reworked it
so it did.

Tests pass without warnings.