Diff schemas against prod
To avoid ending up with the dreaded distributed monolith and keep local
development simple we've been mocking a version of the schema in the frontend
directory. This has allowed the services to be developed independently but it's
easy to have the schema offered by the api and the schema mocked by the
frontend slide out of sync.

This commit adds a step to the CI process where we diff the service's schema
against production. For the frontend, this checks the schema.faker.graphql file
against prod, and for the api it runs `pipenv run schema-export` and diffs the
resulting schema.json against prod.

The thinking is that this is the lowest overhead to get the guarantees we are
looking for.

The schema diff tool just warns currently, it doesn't fail the build. This will
let us test out the idea and see if it works the way we hope.

If this is working well and we want to take a hard nosed approach later we can
add the `--fail` option to the tools arguments.