Manage user state with custom context and hook
This commit adds a UserStateProvider function which makes available a
currentUser object and some helper functions to mutate that state.

The provider is passed an initial state object with the appropriate default
values for the currentUser object:

```javascript
<UserStateProvider initialState={{ userName: null, jwt: null, tfa: null }}>
  <App/>
</UserStateProvider>
```

That inital state is retrieved via the `useUserState()` hook along with helper
functions to operate on that state.

```javascript
const { currentUser, isLoggedIn, login, logout } = useUserState()
```

The `isLoggedIn` function will return true if there are truthy values in the
state (ie, we have non-null values for userName, jwt or tfa).

The `login` function expects the details for a user to be passed in, and will
set the currentUser accordingly:

```javascript
login({ userName: null, jwt: null, tfa: null })
```

The `logout` function works by setting the state back to whatever the initial
state was:
```javascript
logout()
```