Fix lingui plural warning (#293)
This commit started with fixing a warning about plurals not working when the
app was loaded. Fixing that caused some breakage in the tests. Fixing that
revealed that none of the funkyness that broke was needed in the first place so
I'm ripping that out.

The result is everything is much cleaner and there are no warnings to be found
anywhere!