Switch from glob to regex for flux image matching (#305)
For a long time now flux has been confused by the existence of images for our
feature branches. This shouldn't be the case since the only tags it should
consider start with 'master'. This commit switches from the existing glob based
matching to regex in hopes that flux will be better behaved.