Clean up and hide dmarc report page (#315)
This commit does some cleanup of the DmarcReportPage, and removes the links to
it in the app. The idea is to hide it for the moment because it's racing ahead
of the API.

Cleanup-wise, DmarcReportBreakdown has been collapsed into DmarcReportPage
making a pretty massive component. It's clear this needs to get broken down,
but less clear where the boundaries should be.