Move off of Create-React-App
A little while we made the jump from a custom webpack setup to Create React
App. The move was prompted by html-webpack-plugin breaking some of the plugins
we were using.

With a little time behind us, the tradeoffs of this move have become a source
of irritation.

More control over linting (doing it in CI not just when we are running the
app), custom config and direct control over more of our dependencies and build
process are motivating factors in moving back to an updated custom setup.