Split out the Users model & give it responsibility
This commit is the first step in increasing the locality of User related
functionality. The key domain models (Users, Organizations and Affiliations)
have been extracted from the models.py file and given a file of their own to
call home.

This was also the start of moving some logic into the Users model to fight its [anemia](https://martinfowler.com/bliki/AnemicDomainModel.html) and get some
tests going for it.