Refactor auth wrapper tests
This commit reworks the auth wrapper tests with an eye making them fail with
more helpful errors and be more understandable when read.

To improve the understandability, the specific preconditions needed for a given
test (what users and whatnot need to exist for this test to pass) have been
moved inside each test rather than having them created in a single global setup
function.

The acquisition of a token and the creation of an auth header has been folded into
functions to make it clearer what the object under test actually is.

Lastly, Users and other models no longer use specific ids, removing the
possibility of violating uniqueness constraints when tests are run.