Add user to orgs by default (#371)
* Add users to organisations by default

This commit creates an org for each user that is created. This org is the
default org to which ad hoc scans will be attached.

* Add users to an org by default

This commit started it's life as a relatively small feature adding a user to an
org by default to ensure that users has a usable account after logging in, but
ended up with some serious mission creep.

Writing tests for this feature uncovered a lack of isolation between
tests, and tests that mostly failed with `KeyError` or
`'NoneType' object is not subscriptable` errors, as well as test data with
clashing ids.

This change had some far reaching consequences and I basically kept reworking
the tests until the isolation errors went away and I could tell if the "user
orgs" feature had actually broken stuff. If it did, I fixed that too.