Simplify the sign_in_user function
This commit uses the tokenize function and the hybrid roles property that was
added to the Users model and uses them to simplify the sign_in_user function.

With users now getting an organization added by default there is no longer a
"user has no roles" case so this code isn't needed.

Everything else is functionally equivalent, with the sole exception that the
tokenize function generates tokens valid for  1 hour by default.