Require an authenticated user for certain routes
This commit adds a RouteIf component, and makes use of it to provide
"protected" routes on the frontend.

Given that it's the routing equivalent of an if statement the component adopts
some of that terminology:

```javascript
<RouteIf condition={true}
  consequent="/path-if-true"
  alternate="/path-if-false">
<p>child rendered if condition is true</p>
</RouteIf>
```

With that component in place, it's then put to use in the App component.

This modification revealed that the App component was making an unwanted
appearance in the SignInPage tests, so that is removed here too.