Fix timestamp generation
TIL: Python only evaluates default arguments once when the function is being
defined, not once per invocation.

Because of the way this function was written, the jwt "issued at" and expiry
times were being calculated once when the server started.  With token
expiration set for 1 hour, this means we had a 1 hour window before all tokens
would be expired by default.

🤦