Update to Istio 1.6 (#459)
This commit does what it says on the tin, and moves us to Istio 1.6.  As part
of this commit, I've also edited the istio config to remove some extra ports
that istio was listening on. The documentation has been updated to reflect the
extra bit of tomfoolery which we can look to automate or eliminate shortly.