Get certs via http-01 challenge
After some deliberation over the security implications of the dns-01 challenge
vs the http-01 challenge this commit switches us over to the http-01.

The thinking here is that both challenges happen over an unencrypted channel,
and both challenges mitigate MITM attacks by looking at the requester from
multiple network vantage points (http://tiny.cc/vqlypz).

The big tradeoff here is that wildcard certs are only possible via dns-01, at
the cost of a little extra setup... the managing of a DNS hosted zone and a
service account with which to manipulate those settings to complete the
challenge. We weren't/aren't using wildcard certs, and are happy to accept an
operational simplification anywhere that doesn't impact our security posture.

Please see the IETF docs for further discussion of the security considerations:
http://tiny.cc/rrlypz