Flip .dockerignore to default accept (#474)
Previously the .dockerignore file was written to deny everything with
exceptions for the key files required for the app. This is great in theory, but
in practice it leads to semi-regular breakage as the files that constitute the
app change and the .dockerignore does not, resulting in an image that doesn't
work.

The new setup is blocks .env files and some other miscellania but accepts
everything else by default.