Remove calls to app_context (#476)
* Remove calls to app_context

As pointed out in the Flask documentation, "importing the app instance within
the modules in your project is prone to circular import issues".  This commit
removes calls to app_context and the attendant import of the app object.

This will allow us to refactor some of the other parts of the app with fewer
import issues.

The remaining imports of the app object are all in the tests and are all
related to setting headers for testing.

* Group imports a-la PEP

Apparently imports should be grouped in the following order, with a blank line
between each group.

Standard library imports.
Related third party imports.
Local application/library specific imports.