Don't use ci image for frontend builds (#488)
This commit drops the use of the CI image for the frontend. Basically that
image was needed because webpack-dev-server uses chokidar 2, which uses an old
version of fsevents which tries to compile a fs watcher for OSX even on Linux
using python 2 which isn't included in node:alpine.

This commit forces the use of modern version of chokidar and drops the image
entirely.