🍩--
This commit reigns in the number of doughnuts on the landing page and slims
down the code as well.

The nav bar has also been changed to match the title bar in the doughnut box,
and the green used for the "pass" color has been lightened: both of these
changes are being done to reach a better colour contrast for low vision users.