Add Query.findOrganizationDetailBySlug to API
This is a breaking change to the API, since it drops Query.organization and
replaces it with Query.findOrganizationDetailBySlug which returns an
OrganizationDetail type.

The thinking here is to return a specific type that is appropriate for the
public API, meaning that it doesn't expose details that depend on a logged in
user like the base Organization type does. It's also a step away from the very
broad generic names we were using in the early days to narrow specific ones
that take up less "naming real estate" leaving us more room to evolve our API
additively.