Enable zero downtime deploys for the frontend (#567)
This commit bumps the replica count and sets the rolling update parameters to
enable a zero downtime deployment for the frontend. This leverages the already
implemented liveness/readiness checks allowing kubernetes to spin up new
containers and seamlessly redirect traffic to the new containers only when they
are ready to handle it.

This is a key capability allows for constant updates without the risk of
services disruption/downtime for users.