Make URL type more strict
The existing URL type allowed any valid URL, but we don't actually want *all*
URLs. Urls with user credentials in them, mongodb, web sockets, file,
javascript or telephone protocols... there are a bunch of things that are
technically valid but that we really don't want people passing us because they
are dangerous, useless, or we shouldn't/won't/can't scan it.

This is an attempt to tighten our definition of what constitutes a valid url for
us.