Let flux deploy based on timestamps
This commit removes the annotations from the api/frontend deployments which
leaves flux to just deploy whatever is the latest image based on the
timestamps. This wasn't done previously because we were building images for
feature branches as well as the master branch which forced us to be specific.
Since we've stopped building feature branch images we can drop these
annotations.