Pull summaries from the API
This commit deletes the demo data from the code and switches the SummaryGroup
component to pull the data from the API: both doughnuts with a single query.
Also, this commit removes the LandingPage test since it was basically a
duplicate of the SummaryGroup test, displaying the same stuff and adding no new
interesting behaviour.