Upgrade to Apollo Client 3
This commit bumps us to use the new Apollo Client 3. This meant touching a fair
bit of stuff throughout the app given that all their imports had changed. It
seems they've also become more strict about raising errors when there are
differences between the query and response for the MockedProvider. This was/is
annoying, but actually revealed a bunch of problems in our tests, so I'm pretty
happy with that.

One thing to note is that the DomainsPage has some weird stuff going on
requiring a bit of fudging. Basically it now uses a mockLink function that
swallows errors, that are being raised in our test setup due to the way that
the new MockedProvider works. Since this PR is already pretty big, refactoring
the DomainsPage and removing the mockLink will be subject of a follow up PR
reworking that component.