Fix key warning in UserList
This commit adds keys to the Box components being mapped over in the UserList.
This fixes the `Each child in a list should have a unique "key" prop.` warning
that was showing up in the tests.  Also removed a console.log statement which
was also showing up in the test output.