Paginate cache results
So it turns out that Apollo's recommended approach to pagination is to have
their new relayStylePagination helper deal with caching, and that cache simply
returning everything it has leaving it up to the application to decided which
parts of the array of results to display.

This commit adds that "deciding which parts of the array to display" logic to
the usePaginatedCollection hook.