Save codemod script
Codemods are a way of doing language aware edits to large numbers of files.
Given the number of files we are starting to have, and the complexity of some
of the edits we want to make, this is a useful approach.

This particular transform was already applied to the test files in the JS API
folder in 6152cbd1e334fc5, but preserving the learning here will help us with
our next mass refactor.