Refactor to separate platform from app
This is a pretty major refactoring of our Kubernetes config. The idea here is
finally drawing some clean lines between things like Cert Manager which defines
a `Certificate` object and our app, which uses uses it.

The end result is an app folder containing the config for the app, and a
platform folder with the config for the platform.