Run Kube-ArangoDB without HA (#1251)
This commit sets the replica count to 1 for the ArangoDB operator in all the
app environments. The idea is that until the database itself is running HA,
having an HA operator isn't super useful.

This commit also includes a little file shuffling, and a change of some knative
timeout values to ensure the app comes up without errors in all environments.