Update API to lingui v3 (#1489)
* update to v3

* update lingui config

* update config

* update main i18n config

* update translation files

* update remaining files

* fix test outputting console warn

* Remove npx calls from scripts

npm will use locally installed lingui cli by default. calling npx pulls
whatever is the latest version of the cli tool creating a possibility of
version conflicts.

* Get most tests passing

Removing the "missing" key from setupI18n got all the tests passing except for
src/organization/objects/__tests__/organization-summary.test.js

* Remove i18n from organization summary resolvers

This commit removes the translation of "pass" and "fail" from the data since
these are translated on the frontend.

* Keep the data translations

This commit reconsiders the last and lets the api translate the data if
requested.

* Replace bcrypt (C++) with bcryptjs

* Get dotenv-safe working properly with es6

* Use defineMessage to help lingui find strings

The lingui `t` macro is tied to a global object which is problematic for per
request locales. This approach uses defineMessage to create a message that
Lingui knows to extract, and then we can use the `i18n._()` function to
translate it.

* Revert changes to build/start/dev and url-slug

* Update CIPHER_KEY env var in docs

Some of the tests check the length of the key.

* final changes

* add npx to lingui commands

* add extra line in env func file

* Final package fixes

Co-authored-by: nsdeschenes <nsdeschenes@outlook.com>