Move startup messages out of server.js
This commit pulls startup messages out of the server constructor. This change
allows us to start removing some of the mocking of stdout that is going on in
the tests.

This change leaves behind an unused PORT variable (now called _PORT) , which is
retained because it's removal would change the signature of the Server function
causing widespread breakage.

Followup refactorings will be aimed at getting Server to accept an object (to
stop callers from coupling to argument order), and removing stdout mocking
where possible.