Verify jwt token in middleware
This commit adds the [express-jwt
middleware](https://github.com/auth0/express-jwt) to the api creating a central
place to check the validity of tokens.

The middleware gets the incoming request, validates/decodes the jwt and
attaches the result to the request object as request.user.

https://github.com/auth0/express-jwt#retrieving-the-decoded-payload

The request object is already passed into the resolvers via the context so
decisions can be made there based on `request.users`.