Update to the latest version of arango-tools
Arangojs renamed methods on the cursor and transaction objects, which where
returned by Arango-tools. Given code makes some extensive use of those methods,
Arango-tools now returns a proxy object that translates the old calls
(cursor.each and transaction.run) to the new calls (cursor.forEach and
transaction.step).

cursor.hasNext was changed from a function to a boolean so this was fixed by a
simple search and replace.

Details:
https://arangodb.github.io/arangojs/CHANGELOG#cursor-api-1