Update to istio 1.7.7 (#1642)
This commit bumps us up to the latest istio that knative supports. This update
comes with a little config shuffling to support kiali and prometheus as addons.
Where previous updates have required manual editing of the IngressGateway
service, this commit finally makes full use of kustomizes patchStrategicMerge
option and uses it to delete the offending ports. No more manual edits.