Get superadmin job working
This commit finally makes it easier to use the superadmin service.
A superuser .env is created by the makefile, and the job definition now uses
the credentials that are in the secret that .env file creates.
Both the job and the secret are in the superadmin namespace.

`make secrets` now accepts parameters that set the values of the superadmin.env.

The following values are the default and are optional, but specifying them
obviously supplies better values.

```
make credentials mode=dev displayname=admin email=admin@example.com password=admin
```

This means that setting up an instance for the first time, you have an easy way
to get a base account going.