Add one time backup job
This commit adds a one time backup job.

The makefile has been changed to create a backup job in the db namespace.
The backup job uses `arangodump` and `rclone` create a backup and upload it to
either Google Cloud Storage or Azure Blob Storage.

The backup job can be triggered like so:
```
make backup env=gke
```