Use arango middleware (#2680)
This commit adds middleware which attaches the query and transaction functions
to the express request object.

This allows us to get at the query function without needing to pass it around
separately. It also helps our observability by passing the
x-cloud-trace-context header along with database queries allowing for end to
end tracing.