Let affiliations fail and display the rest
Organization details includes the org tombstone stuff, a list of domains, and a list of affiliated users.

Normal users don't have permission to list users, so the query returns the org tombstone, the domains, affiliations set to "null" and an error: "Cannot query affiliations on organization without admin permission or higher."

This error causes the entire response to the thrown away.

This commit uses an error policy of "ignore" to ignore the permissions error and use whatever was returned.