Refactor guidance service
This refactor simplifies the guidance service, dropping the api calls to the
github api in favour of reading the guidance directly from the filesystem.
This allows us to drop the credentials needed for those api calls as well.

Somewhat controversially, tests have been deleted owning to the fact that the
changes were significant, and this entire service will be rewritten in
JavaScript shortly.