Add timestamps to edges before saving them
This commit adds timestamps to the various edges so that graph traversals can
be constrained by time. For example, a traversal that gets the latest https
scan result for a domain.
Currently this is harder than it needs to be because timestamps live on the
scan result, rather than the edge that connects it to a domain.