Add failure analysis scripts
Failing people for stuff that is explicitly marked as "phase out" or "weak" by
the Cybercentre is pretty uncontroversial. However there are many cases where
we fail people for using things that aren't explicitly approved by the
Cybercentre. This is a different ball of wax that we need to take stock of.
These scripts do that job: generating a summary of cipher and curve failures
that are *not* on the Cybercentre's naughty list. This is both a sanity check
for Tracker itself and also potentially helpful information for the Cybercentre
so they can prioritize their cryptographic work.