Add Codemod to upgrade arango-tools
The latest version of arango-tools has been changed to make it viable in an
initContainer. Adopting this version requires a complex edit to 120 files, so
this commit adds a codemod to do that work.

The codemod is run with the following command in the codemods folder:
```
npm run transform -- -t upgrade-at.js ../../api/src/ --schemaPath=/home/mike/projects/tracker/api/database.json
```