Refactor database handling.
Currently the API attempts to "migrate" the database when it starts. This
commit is the first step in moving this behaviour into an initContainer where
it belongs.
This commit removes the use of arango-tools at service startup... and reworks
the tests to use the new initContainer friendly args to arango-tools in the
tests.