Fix dotenv import
This commit addresses a weird bug: starting the api with `npm run dev` would
result in the notify client throwing an error.

This ends up being due to how import/require are evaluated, and the previous
ordering meant that our .env wasn't actually added to the environment.

Details in motdotla/dotenv/#509

Since we use dotenv via dotenv-safe... we never saw that.