Rework istio-ingressgateway config (#3669)
This commit puts the istio-controlplane config in the namespace that the istio
operator is watching. It also reworks the way the manifests are created, making
sure the operator is separate from the rest of the istio config (it's included
both in `istioctl manifests generate` and `istioctl operator dump`) and can be
updated and included on it's own.

The end result is we should have 2 ingressgateways and 2 istiods. Again.